+++ src/pyquanda/ansible/__init__.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """Main Ansible Module Builder."""
-# pylint: disable=unused-argument
+# pylint: disable=unused-argument,missing-function-docstring
 
+import json
 import os
 import shutil
 import tempfile
@@ -14,7 +15,7 @@ from ruamel.yaml import YAML
 
 from pyquanda.exceptions import PreCheckFail
 from pyquanda.hooks.registry import HookLoader
-from pyquanda.hooks.config import EMITTER
+from pyquanda.hooks.config import SYNC_EMITTER
 from pyquanda.hooks.config import (
     HOOK_TYPE_ANSIBLE_EVENT,
     HOOK_TYPE_ANSIBLE_RUNBOOK_COMPLETE,
@@ -25,7 +26,7 @@ _POP_KEYS = ["env", "ansible_facts"]
 _yaml = YAML()
 
 _BASE = Path(__file__).parent
-_SITE_YAML_FILE = _BASE.joinpath("site.yaml")
+_SITE_PLAY_FILE = _BASE.joinpath("play.yaml")
 
 SUPRESS_OUTPUT = False
 
@@ -37,7 +38,6 @@ localhost
 
 EVENT_FAILURE = "runner_on_failed"
 EVENT_OK = "runner_on_ok"
-EVENT_RUNBOOK_DONE = "successful"
 
 
 class Ansible:
@@ -51,72 +51,46 @@ class Ansible:
         DIR_HANDLER,
         DIR_TASKS,
     ]
+    PYTHON_INTERP = shutil.which("python3")
 
-    def __init__(self, path: Path) -> None:
-        """__init__.
-
-        Args:
-            path (Path): path
-
-        Returns:
-            None:
-        """
-        HookLoader.load()
-        self._verify(path)
-        self.path = path
-        self.name = self.path.name.title()
-        self.playbook = list(_yaml.load(_SITE_YAML_FILE.read_bytes()))
-        self.playbook[0]["name"] = self.name
-        self.python_interp = shutil.which("python3")
-
-    def status_handler(self, args: Dict, **kwargs):  # type: ignore
-        """status_handler.
-
-        Args:
-            args: ansible event args
-            kwargs: Not used
-        """
-        # gd linter
-        _ = kwargs
-        status = args["status"]
-        if status == EVENT_RUNBOOK_DONE:
-            send_d = {
-                "runbook": self.name,
-                "status": "complete",
-            }
-            EMITTER.emit(HOOK_TYPE_ANSIBLE_RUNBOOK_COMPLETE, send_d)
-        # print(f"runbook status:{status}, name:{self.name}")
-
-    def event_handler(self, evt_dct: Dict):
-        """event_handler.
-
-        Args:
-            evt_dct (Dict): evt_dct
-        """
-        respond = ["runner_on_ok", "runner_on_failed"]
+    @classmethod
+    def event_handler(cls, evt_dct: Dict) -> bool:
+        respond = [
+            EVENT_FAILURE,
+            # EVENT_OK,
+        ]
         evt = evt_dct["event"]
         if evt not in respond:
-            return
+            return True
+
         data = evt_dct.get("event_data", {})
-        res = data.get("res", {})
+        ignore_errors = data.get("ignore_errors", False)
+        if ignore_errors:
+            return True
+
+        if "res" in data:
+            res = data.pop("res")
+        else:
+            res = {}
         for key in _POP_KEYS:
             if key in data:
                 data.pop(key)
             if key in res:
                 res.pop(key)
         send_d = {
-            "runbook": self.name,
             "event": evt,
+            "data": data,
             "res": res,
         }
-        EMITTER.emit(HOOK_TYPE_ANSIBLE_EVENT, send_d)
-        # print(f"{evt}", file=sys.stdout)
-        #  if evt == "runner_on_failed":
-        #      print(
-        #          "event",
-        #          json.dumps(evt_dct, indent=4, separators=(",", " : ")),
-        #          file=sys.stderr,
-        #      )
+        # for debugging later
+        #  import json
+        #
+        #  print("*" * 100)
+        #  print(ignore_errors)
+        #  print(json.dumps(send_d, indent=4, separators=(",", " : ")))
+        #  print("*" * 100)
+        SYNC_EMITTER.emit(HOOK_TYPE_ANSIBLE_EVENT, send_d)
+        return True
 
     @classmethod
     def copy_dirs(cls) -> List[str]:
@@ -129,8 +103,9 @@ class Ansible:
         """
         return [getattr(cls, i) for i in dir(cls) if i.startswith("DIR_")]
 
+    @classmethod
     def _copy_src_role_to_dest_dir(
-        self, src_mod_path: Path, dst_roles_dir: Path
+        cls, src_mod_path: Path, dst_roles_dir: Path
     ):
         """_copy_src_role_to_dest_dir.
 
@@ -148,9 +123,8 @@ class Ansible:
             shutil.copy(src_config_file, dst_config_file)
         except FileNotFoundError:
             pass
-        self.playbook[0]["roles"].append(str(bname))
         dst_dir = dst_roles_dir.absolute().joinpath(bname)
-        for req_dir in self.copy_dirs():
+        for req_dir in cls.copy_dirs():
             to_dir = dst_dir.joinpath(req_dir)
             src_dir = src_mod_path.joinpath(req_dir)
             try:
@@ -167,46 +141,150 @@ class Ansible:
                 )
                 raise PreCheckFail(msg) from _e
 
-    def run(self):
+    @classmethod
+    def run_single(cls, module_path: Path, debug_only: bool):
         """run."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            tdir = Path(tmpdir)
-            role_dir = tdir.joinpath("roles")
-            role_dir.mkdir()
-            hosts_file = tdir.joinpath("hosts")
-            hosts_file.write_text(_HOSTS)
-            self._copy_src_role_to_dest_dir(self.path, role_dir)
-            pb_file = tdir.joinpath("site.yaml")
-            cfg = tdir.joinpath("ansible.cfg")
-            cfg.write_text(
-                "\n".join(
-                    [
-                        "[defaults]",
-                        f"interpreter_python={self.python_interp}",
-                    ]
-                )
-            )
-            os.environ["ANSIBLE_CONFIG"] = str(cfg)
-            with pb_file.open("wb") as fileh:
-                _yaml.dump(self.playbook, fileh)
-            # run status for the kickoff TODO
-            syspath = os.environ["PATH"]
-            syspath += ":/usr/local/bin"
-            inv = {"all": {"hosts": {"localhost": None}}}
-            ansible_runner.run(
-                event_handler=self.event_handler,
-                status_handler=self.status_handler,
-                settings={
-                    "suppress_ansible_output": SUPRESS_OUTPUT,
-                },
-                inventory=inv,
-                playbook=str(pb_file),
-                envvars={
-                    "PATH": syspath,
-                },
-            )
+        cls._run(module_path, is_all=False, debug_only=debug_only)
 
-    def _verify(self, path: Path):
+    @classmethod
+    def _get_sorted_module_paths(cls, paths: List[Path]) -> List[Path]:
+        def _sort(path):
+            cfg = path.joinpath("config.yaml")
+            obj = _yaml.load(cfg)
+            return obj["order"]
+
+        errors = []
+        mod_paths = []  # type: List[Path]
+        for i in sorted(paths, key=_sort):
+            try:
+                cls._verify(i)
+                mod_paths.append(i)
+            except PreCheckFail as _e:
+                errors.append(_e)
+        if errors:
+            msg = ",".join(str(i) for i in errors)
+            raise PreCheckFail(msg)
+        return mod_paths
+
+    @classmethod
+    def _get_playbook(cls, path: Path):
+        var_file = path.joinpath("vars.yaml")
+        config_file = path.joinpath("config.yaml")
+
+        cfg_obj = _yaml.load(config_file)  # type: Dict
+        if var_file.exists():
+            vobj = _yaml.load(var_file)
+            cfg_obj.update(vobj)
+
+        pb = _yaml.load(_SITE_PLAY_FILE.read_bytes())  # type: Dict
+        pb["name"] = path.name.title()
+        task = {
+            "block": [
+                {
+                    "include_role": {
+                        "name": path.name,
+                    },
+                    "vars": cfg_obj,
+                },
+            ],
+            "rescue": [
+                {
+                    "debug": {
+                        "msg": "Play failed",
+                    },
+                },
+                {
+                    "meta": "clear_host_errors",
+                },
+                {
+                    "meta": "end_play",
+                },
+            ],
+        }
+        pb["tasks"].append(task)
+        # print(json.dumps(pb, indent=4, separators=(",", " : ")))
+        return pb
+
+    @classmethod
+    def _run(cls, path: Path, is_all: bool, debug_only: bool):
+        """run."""
+        HookLoader.load()
+        mod_paths = []
+        errors = []
+        playbooks = []
+
+        if is_all:
+            mvals = [i for i in path.iterdir() if i.is_dir()]
+            mod_paths = cls._get_sorted_module_paths(mvals)
+        else:
+            mod_paths = cls._get_sorted_module_paths([path])
+
+        if errors:
+            msg = ",".join(str(i) for i in errors)
+            raise PreCheckFail(msg)
+
+        for i in mod_paths:
+            playbooks.append(cls._get_playbook(i))
+
+        if debug_only:
+            print(json.dumps(playbooks, indent=4, separators=(",", " : ")))
+        else:
+            with tempfile.TemporaryDirectory() as tmpdir:
+                tdir = Path(tmpdir)
+                role_dir = tdir.joinpath("roles")
+                role_dir.mkdir()
+                hosts_file = tdir.joinpath("hosts")
+                hosts_file.write_text(_HOSTS)
+                for i in mod_paths:
+                    cls._copy_src_role_to_dest_dir(i, role_dir)
+
+                cfg = tdir.joinpath("ansible.cfg")
+                cfg.write_text(
+                    "\n".join(
+                        [
+                            "[defaults]",
+                            f"interpreter_python={cls.PYTHON_INTERP}",
+                        ]
+                    )
+                )
+                os.environ["ANSIBLE_CONFIG"] = str(cfg)
+                pb_file = tdir.joinpath("site.yaml")
+                with pb_file.open("wb") as fileh:
+                    _yaml.dump(playbooks, fileh)
+                # run status for the kickoff TODO
+                syspath = os.environ["PATH"]
+                syspath += ":/usr/local/bin"
+                inv = {"all": {"hosts": {"localhost": None}}}
+                run = ansible_runner.run(
+                    event_handler=cls.event_handler,
+                    # status_handler=cls.status_handler,
+                    settings={
+                        "suppress_ansible_output": SUPRESS_OUTPUT,
+                    },
+                    inventory=inv,
+                    playbook=str(pb_file),
+                    envvars={
+                        "PATH": syspath,
+                    },
+                )
+                stats = {
+                    k: 0 if not v.values() else list(v.values())[0]
+                    for k, v in run.stats.items()  # type: ignore
+                }
+                send_d = {
+                    "status": run.status,  # type: ignore
+                    "rc": run.rc,  # type: ignore
+                    "stats": stats,
+                }
+                SYNC_EMITTER.emit(HOOK_TYPE_ANSIBLE_RUNBOOK_COMPLETE, send_d)
+
+    @classmethod
+    def run_all(cls, modules_directory: Path, debug_only: bool):
+        """run."""
+        cls._run(modules_directory, is_all=True, debug_only=debug_only)
+
+    @classmethod
+    def _verify(cls, path: Path):
         """_verify.
 
         Args:
@@ -215,13 +293,13 @@ class Ansible:
         Raises:
             PreCheckFail: on validation error
         """
-        for i in self.copy_dirs():
+        for i in cls.copy_dirs():
             dpath = path.joinpath(i)
             if not dpath.exists():
                 raise PreCheckFail(f"Path {dpath} does not exist")
             if not dpath.is_dir():
                 raise PreCheckFail(f"Path {dpath} is not a directory")
-        for i in self.MK_MAINS:
+        for i in cls.MK_MAINS:
             mfile = path.joinpath(i, "main.yaml")
             if not mfile.exists():
                 raise PreCheckFail(f"file {i} does not exist")